+++ spring_exercise/src/main/java/at/bbrz/spring_exercise/controller/model/LoginChecker.java
@@ -2,5 +2,5 @@ package at.bbrz.spring_exercise.controller.model;
 
 public interface LoginChecker {
     boolean check(LoginUser loginUser);
-    int getUserId(LoginUser loginUser);
+    Long getUserId(LoginUser loginUser);
 }
